OP-72 Change all the double type in Greedy Scheduler to int

+++ src/main/java/op/model/ScheduledTask.java
@@ -25,7 +25,7 @@ public class ScheduledTask {
         return processor;
     }
 
-	public double getStartTime() {
+	public int getStartTime() {
 		return startTime;
 	}
